[Refactor]: 제네릭 타입 변경

+++ src/main/java/hongik/heavyYoung/domain/item/repository/ItemCategoryRepository.java
@@ -3,5 +3,5 @@ package hongik.heavyYoung.domain.item.repository;
 import hongik.heavyYoung.domain.item.entity.ItemCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemCategoryRepository extends JpaRepository<ItemCategory, Integer> {
+public interface ItemCategoryRepository extends JpaRepository<ItemCategory, Long> {
 }
